CA-290644: vm-resume: upgrade profile from qemu-trad to qemu-upstream-compat
in the destination host only when resuming a VM that has been started using a
previous version of xenopsd that has persistent metadata version 0.

This will usually apply to a cross-version vm-resume from a host using an
earlier version of xenopsd or RPU from a slave to a master.

This will not apply to VMs with qemu-trad profile resuming in a pool where
they were started with a persistent metadata version 1 or later.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>